requirements.txt
fastapi>=0.135.0
uvicorn==0.24.0
qrcode[pil]==7.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx>=0.28.0
apscheduler==3.10.4
PyMuPDF==1.25.1
reportlab==4.1.0
google-genai>=1.0.0
